+++ lua/onedark/theme.lua
@@ -4,14 +4,6 @@ local config_module = require("onedark.config")
 
 local M = {}
 
-local function apply_overrides(group, overrides)
-  for k, v in pairs(overrides) do
-    if group[k] ~= nil and type(v) == "table"  then
-      group[k] = v
-    end
-  end
-end
-
 ---@param config onedark.Config
 ---@return onedark.Theme
 function M.setup(config)
@@ -594,13 +586,6 @@ function M.setup(config)
     CocUnderline = {style = "undercurl"}
   }
 
-  theme.defer = {}
-
-  local overrides = config.overrides(c)
-  apply_overrides(theme.base, overrides)
-  apply_overrides(theme.plugins, overrides)
-  apply_overrides(theme.defer, overrides)
-
   if config.hide_inactive_statusline then
 
     -- StatusLine
@@ -615,6 +600,10 @@ function M.setup(config)
       theme.base.StatusLineNC = inactive
     end
   end
+
+  local overrides = config.overrides(c)
+  util.apply_overrides(theme.base, overrides)
+  util.apply_overrides(theme.plugins, overrides)
 
   return theme
 end